64-bit clean

+++ hx.cc
@@ -494,7 +494,7 @@
 			out,
 			"<span class=\"%s\">%.*s</span>",
 			cls,
-			s.size(),
+			(int) s.size(),
 			s.data()
 		);
 	}
@@ -1758,7 +1758,7 @@
 				status.codeName +
 					sizeof(status.codeName),
 			" [%c], [%.*s]", status.codeSpecial,
-			sizeof(status.codeName), status.codeName
+			(int) sizeof(status.codeName), status.codeName
 		);
 		*status.codeNameEnd++ = ch;
 		continue;
